+++ microgradx/optim/__init__.py
@@ -1,5 +1,6 @@
 from microgradx.optim.optimizer import Optimizer, clip_grad_norm_, clip_grad_value_
 from microgradx.optim.sgd import SGD
+from microgradx.optim.adam import Adam
 from microgradx.optim.adamw import AdamW
 from microgradx.optim.lion import Lion
 from microgradx.optim.lr_scheduler import (
@@ -10,9 +11,11 @@ from microgradx.optim.lr_scheduler import (
     CosineAnnealingLR,
     LinearWarmup,
     OneCycleLR,
+    ReduceLROnPlateau,
 )
 
-__all__ = ["Optimizer", "SGD", "AdamW", "Lion",
+__all__ = ["Optimizer", "SGD", "Adam", "AdamW", "Lion",
            "clip_grad_norm_", "clip_grad_value_",
            "LambdaLR", "StepLR", "MultiStepLR", "ExponentialLR",
-           "CosineAnnealingLR", "LinearWarmup", "OneCycleLR"]
+           "CosineAnnealingLR", "LinearWarmup", "OneCycleLR",
+           "ReduceLROnPlateau"]